requirements.txt
# Core app
streamlit==1.37.1
pyyaml==6.0.2
python-dotenv==1.0.1

# Retrieval
chromadb==0.5.5
sentence-transformers==3.0.1

# Local models (CPU okay; GPU if available)
torch>=2.1.0
transformers==4.44.0
accelerate==0.33.0

# Language utilities
langdetect==1.0.9

# Optional eval
ragas==0.1.14
datasets==2.21.0
evaluate==0.4.2
